Allow setting an expiration time for unauthenticated notification events (#169)
* Allow setting an expiration time for unauthenticated notification events

The unauthenticated notifications proxy gives access to event data in
perpetuity. If a URL is shared or leaks, it is valid for the lifetime of
the event id.

This change allows the user to set an expiration time (in minutes) for
the notification event to be valid. After this time, the notification
proxy will not allow the event to be accessed, and the user will have to
go through the typical Frigate flow to retrieve snapshots/clips etc.

If the user sets the expiration time to 0 (default) this feature is
disabled.

* Applying PR comments

* Adapt testing to allow unauthenticated client tests.

* Use seconds instead of minutes.

* Minor naming improvement.

* Rename translation.

Co-authored-by: Dermot Duffy <dermot.duffy@gmail.com>